+++ core/src/main/java/org/vertexium/query/DefaultGraphQueryIterableWithAggregations.java
@@ -1,7 +1,6 @@
 package org.vertexium.query;
 
-import org.vertexium.VertexiumException;
-import org.vertexium.VertexiumObject;
+import org.vertexium.*;
 
 import java.util.*;
 import java.util.stream.Collectors;
@@ -38,6 +37,9 @@ public class DefaultGraphQueryIterableWithAggregations<T extends VertexiumObject
         if (agg instanceof CalendarFieldAggregation) {
             return true;
         }
+        if (agg instanceof CardinalityAggregation) {
+            return true;
+        }
         return false;
     }
 
@@ -49,7 +51,35 @@ public class DefaultGraphQueryIterableWithAggregations<T extends VertexiumObject
         if (agg instanceof CalendarFieldAggregation) {
             return (TResult) getCalendarFieldHistogramResult((CalendarFieldAggregation) agg, it);
         }
+        if (agg instanceof CardinalityAggregation) {
+            return (TResult) getCardinalityAggregationResult((CardinalityAggregation) agg, it);
+        }
         throw new VertexiumException("Unhandled aggregation: " + agg.getClass().getName());
+    }
+
+    private CardinalityResult getCardinalityAggregationResult(CardinalityAggregation agg, Iterator<T> it) {
+        String fieldName = agg.getPropertyName();
+
+        if (Element.ID_PROPERTY_NAME.equals(fieldName)
+                || Edge.LABEL_PROPERTY_NAME.equals(fieldName)
+                || Edge.OUT_VERTEX_ID_PROPERTY_NAME.equals(fieldName)
+                || Edge.IN_VERTEX_ID_PROPERTY_NAME.equals(fieldName)
+                || ExtendedDataRow.TABLE_NAME.equals(fieldName)
+                || ExtendedDataRow.ROW_ID.equals(fieldName)
+                || ExtendedDataRow.ELEMENT_ID.equals(fieldName)
+                || ExtendedDataRow.ELEMENT_TYPE.equals(fieldName)) {
+            Set<Object> values = new HashSet<>();
+            while (it.hasNext()) {
+                T vertexiumObject = it.next();
+                Iterable<Object> propertyValues = vertexiumObject.getPropertyValues(fieldName);
+                for (Object propertyValue : propertyValues) {
+                    values.add(propertyValue);
+                }
+            }
+            return new CardinalityResult(values.size());
+        } else {
+            throw new VertexiumException("Cannot use cardinality aggregation on properties with visibility: " + fieldName);
+        }
     }
 
     private TermsResult getTermsAggregationResult(TermsAggregation agg, Iterator<T> it) {